selfservice-api/requirements/dev.txt
# Everything the developer needs in addition to the production requirements
-r prod.txt

# Testing
pytest
pytest-mock
pytest-cov
requests
pyhamcrest

# Lint and code style
flake8
flake8-blind-except
flake8-debugger
flake8-docstrings
flake8-isort
flake8-quotes
pep8-naming
autopep8
pylint
pylint-flask